+++ internal/management/controller/database_controller_test.go
@@ -25,7 +25,6 @@ import (
 	corev1 "k8s.io/api/core/v1"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 	"k8s.io/apimachinery/pkg/types"
-	"k8s.io/utils/ptr"
 	ctrl "sigs.k8s.io/controller-runtime"
 	"sigs.k8s.io/controller-runtime/pkg/client"
 	"sigs.k8s.io/controller-runtime/pkg/client/fake"
@@ -46,153 +45,6 @@ type fakeInstanceData struct {
 func (f *fakeInstanceData) GetSuperUserDB() (*sql.DB, error) {
 	return f.db, nil
 }
-
-var _ = Describe("Managed Database SQL", func() {
-	var (
-		dbMock   sqlmock.Sqlmock
-		db       *sql.DB
-		database *apiv1.Database
-		r        *DatabaseReconciler
-		err      error
-	)
-
-	BeforeEach(func() {
-		db, dbMock, err = sqlmock.New(sqlmock.QueryMatcherOption(sqlmock.QueryMatcherEqual))
-		Expect(err).ToNot(HaveOccurred())
-
-		database = &apiv1.Database{
-			ObjectMeta: metav1.ObjectMeta{
-				Name: "db-one",
-			},
-			Spec: apiv1.DatabaseSpec{
-				ClusterRef: corev1.LocalObjectReference{
-					Name: "cluster-example",
-				},
-				Name:  "db-one",
-				Owner: "app",
-			},
-		}
-	})
-
-	AfterEach(func() {
-		Expect(dbMock.ExpectationsWereMet()).To(Succeed())
-	})
-
-	Context("detectDatabase", func() {
-		It("returns true when it detects an existing Database", func(ctx SpecContext) {
-			expectedValue := sqlmock.NewRows([]string{""}).AddRow("1")
-			dbMock.ExpectQuery(`SELECT count(*)
-		FROM pg_database
-		WHERE datname = $1`).WithArgs(database.Spec.Name).WillReturnRows(expectedValue)
-
-			dbExists, err := r.detectDatabase(ctx, db, database)
-			Expect(err).ToNot(HaveOccurred())
-			Expect(dbExists).To(BeTrue())
-		})
-
-		It("returns false when a Database is missing", func(ctx SpecContext) {
-			expectedValue := sqlmock.NewRows([]string{""}).AddRow("0")
-			dbMock.ExpectQuery(`SELECT count(*)
-		FROM pg_database
-		WHERE datname = $1`).WithArgs(database.Spec.Name).WillReturnRows(expectedValue)
-
-			dbExists, err := r.detectDatabase(ctx, db, database)
-			Expect(err).ToNot(HaveOccurred())
-			Expect(dbExists).To(BeFalse())
-		})
-	})
-
-	Context("createDatabase", func() {
-		It("should create a new Database", func(ctx SpecContext) {
-			database.Spec.IsTemplate = ptr.To(true)
-			database.Spec.Tablespace = "myTablespace"
-			database.Spec.AllowConnections = ptr.To(true)
-			database.Spec.ConnectionLimit = ptr.To(-1)
-
-			expectedValue := sqlmock.NewResult(0, 1)
-			expectedQuery := fmt.Sprintf(
-				"CREATE DATABASE %s IS_TEMPLATE %v OWNER %s TABLESPACE %s "+
-					"ALLOW_CONNECTIONS %v CONNECTION LIMIT %v",
-				pgx.Identifier{database.Spec.Name}.Sanitize(), *database.Spec.IsTemplate,
-				pgx.Identifier{database.Spec.Owner}.Sanitize(), pgx.Identifier{database.Spec.Tablespace}.Sanitize(),
-				*database.Spec.AllowConnections, *database.Spec.ConnectionLimit,
-			)
-			dbMock.ExpectExec(expectedQuery).WillReturnResult(expectedValue)
-
-			err = r.createDatabase(ctx, db, database)
-			Expect(err).ToNot(HaveOccurred())
-		})
-	})
-
-	Context("patchDatabase", func() {
-		It("should reconcile an existing Database", func(ctx SpecContext) {
-			database.Spec.Owner = "newOwner"
-			database.Spec.IsTemplate = ptr.To(true)
-			database.Spec.AllowConnections = ptr.To(true)
-			database.Spec.ConnectionLimit = ptr.To(-1)
-			database.Spec.Tablespace = "newTablespace"
-
-			expectedValue := sqlmock.NewResult(0, 1)
-
-			// Mock Owner DDL
-			ownerExpectedQuery := fmt.Sprintf(
-				"ALTER DATABASE %s OWNER TO %s",
-				pgx.Identifier{database.Spec.Name}.Sanitize(),
-				pgx.Identifier{database.Spec.Owner}.Sanitize(),
-			)
-			dbMock.ExpectExec(ownerExpectedQuery).WillReturnResult(expectedValue)
-
-			// Mock IsTemplate DDL
-			isTemplateExpectedQuery := fmt.Sprintf(
-				"ALTER DATABASE %s WITH IS_TEMPLATE %v",
-				pgx.Identifier{database.Spec.Name}.Sanitize(),
-				*database.Spec.IsTemplate,
-			)
-			dbMock.ExpectExec(isTemplateExpectedQuery).WillReturnResult(expectedValue)
-
-			// Mock AllowConnections DDL
-			allowConnectionsExpectedQuery := fmt.Sprintf(
-				"ALTER DATABASE %s WITH ALLOW_CONNECTIONS %v",
-				pgx.Identifier{database.Spec.Name}.Sanitize(),
-				*database.Spec.AllowConnections,
-			)
-			dbMock.ExpectExec(allowConnectionsExpectedQuery).WillReturnResult(expectedValue)
-
-			// Mock ConnectionLimit DDL
-			connectionLimitExpectedQuery := fmt.Sprintf(
-				"ALTER DATABASE %s WITH CONNECTION LIMIT %v",
-				pgx.Identifier{database.Spec.Name}.Sanitize(),
-				*database.Spec.ConnectionLimit,
-			)
-			dbMock.ExpectExec(connectionLimitExpectedQuery).WillReturnResult(expectedValue)
-
-			// Mock Tablespace DDL
-			tablespaceExpectedQuery := fmt.Sprintf(
-				"ALTER DATABASE %s SET TABLESPACE %s",
-				pgx.Identifier{database.Spec.Name}.Sanitize(),
-				pgx.Identifier{database.Spec.Tablespace}.Sanitize(),
-			)
-			dbMock.ExpectExec(tablespaceExpectedQuery).WillReturnResult(expectedValue)
-
-			err = r.patchDatabase(ctx, db, database)
-			Expect(err).ToNot(HaveOccurred())
-		})
-	})
-
-	Context("dropDatabase", func() {
-		It("should drop an existing Database", func(ctx SpecContext) {
-			expectedValue := sqlmock.NewResult(0, 1)
-			expectedQuery := fmt.Sprintf(
-				"DROP DATABASE IF EXISTS %s",
-				pgx.Identifier{database.Spec.Name}.Sanitize(),
-			)
-			dbMock.ExpectExec(expectedQuery).WillReturnResult(expectedValue)
-
-			err = r.dropDatabase(ctx, db, database)
-			Expect(err).ToNot(HaveOccurred())
-		})
-	})
-})
 
 var _ = Describe("Managed Database status", func() {
 	var (
